math/src/tuple.rs
#[derive(Debug, PartialEq, Copy, Clone)]
pub struct Tuple {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

impl Tuple {
    pub fn new(x: f32, y: f32, z: f32, w: f32) -> Tuple {
        Tuple { x, y, z, w }
    }

    pub fn point(x: f32, y: f32, z: f32) -> Tuple {
        Tuple { x, y, z, w: 1.0 }
    }

    pub fn vector(x: f32, y: f32, z: f32) -> Tuple {
        Tuple { x, y, z, w: 0.0 }
    }

    pub fn is_point(&self) -> bool {
        self.w == 1.0
    }

    pub fn is_vector(&self) -> bool {
        self.w == 0.0
    }
}

impl std::ops::Add for Tuple {
    type Output = Tuple;

    fn add(self, t: Tuple) -> Tuple {
        Tuple::new(self.x + t.x, self.y + t.y, self.z + t.z, self.w + t.w)
    }
}

#[cfg(test)]
#[path = "./tuple_test.rs"]
mod tests;
